% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_helpers.R
\name{modify_query}
\alias{modify_query}
\title{Modify query based on table-specific logic}
\usage{
modify_query(query, table, can_view_others, private)
}
\arguments{
\item{query}{The base dplyr query object.}

\item{table}{The table name.}

\item{can_view_others}{Boolean indicating if the user can view others' data.}

\item{private}{The private list-like object containing user and project info.}
}
\value{
A modified dplyr query object.
}
\description{
Modify query based on table-specific logic
}
\keyword{internal}
